% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitlmm.R
\name{get_BLUP}
\alias{get_BLUP}
\title{\code{Fit MLM to get the BLUP values}}
\usage{
get_BLUP(data = df, model = Brix ~ (1 | Line) + (1 | Loc) + (1 | Year) + (1
  | Line:Loc) + (1 | Line:Year), which.factor = "Line",
  outfile = "data/blup.csv")
}
\arguments{
\item{data}{Phenotypic data [data.frame], see example: "data/pheno.csv".}

\item{model}{[model], find help from lme4 function lmer.}

\item{which.factor}{The factor used for BLUP calculation, normally Genotype or Line [character].}

\item{outfile}{Output file name [character="blup.csv"].}
}
\value{
return fitted model.
}
\description{
Get best linear unbiased predictors for phenotypic values.
Require package "lme4".
}
\examples{
pheno <-  read.csv("data/pheno.csv", header=T)
fit <- get_BLUP(data = df, model = Brix ~ (1|Line) + (1|Loc) + (1|Year) + (1|Line:Loc) + (1|Line:Year),
                 which.factor = "Line", outfile="data/blup.csv")

}

