% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitlmm.R
\name{get_H2}
\alias{get_H2}
\title{\code{Get Broad sense heritability from BLUP model.}}
\usage{
get_H2(fit, numerator = "Line", denominator = data.frame(f = c("Line",
  "Line:Year", "Line:Loc", "Residual"), df = c(1, 2, 2, 4)))
}
\arguments{
\item{fit}{Model fit returned from get_BLUP [model].}

\item{numerator}{Numerator of Vg/Ve, [char].}

\item{denominator}{Denominator of Vg/Ve [data.table], 2nd, degree of freedom.
i.e. data.frame(f=c("Line", "Line:Year", "Line:Loc", "Residual"), df=c(1, 2, 2, 4))}
}
\value{
return A value of heritability.
}
\description{
Run the model fit using function "get_BLUP" and get H2.
}
\examples{
pheno <-  read.csv("data/pheno.csv", header=T)
fit <- get_BLUP(data = df, model = Brix ~ (1|Line) + (1|Loc) + (1|Year) + (1|Line:Loc) + (1|Line:Year),
                 which.factor = "Line", outfile="data/blup.csv")
get_H2(fit, numerator="Line",
       denominator=data.frame(f=c("Line", "Line:Year", "Line:Loc", "Residual"),
       df=c(1, 2, 2, 4)))

}

