% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpheno.R
\name{sim_qtl_pheno}
\alias{sim_qtl_pheno}
\title{\code{Simulate QTL phenotype with normal distribution}}
\usage{
sim_qtl_pheno(geno, h2, alpha, nqtl, distribution = "norm")
}
\arguments{
\item{geno}{genotype data, col1=genoid, col2 and after=snpid, coding: 0,1,2 (no missing data allowed) [data.frame].}

\item{h2}{Broad sense heritability of the trait [numeric(0-1)].}

\item{alpha}{Alpha.}

\item{nqtl}{number of QTL [interger].}

\item{distribution}{[character=norm]}
}
\value{
return A list of many values.
}
\description{
Simulate phenotype using genotype data given number of QTLs and heritability.
}
\examples{
geno <- read.table("data/geno.txt", header=TRUE)
pheno <- sim_qtl_pheno(geno, h2=0.7, alpha=0.5, nqtl=10, distribution="norm")
y <- pheno[['y']]

}
